+++ services/importer/spec/unit/kml_splitter_spec.rb
@@ -1,7 +1,7 @@
 # encoding: utf-8
-gem 'minitest'
-require 'minitest/autorun'
-require_relative '../../lib/importer/kml_splitter'
+
+# Unp includes reference to kml_splitter
+require_relative '../../lib/importer/unp'
 
 include CartoDB::Importer2
 
@@ -15,21 +15,21 @@ describe KmlSplitter do
   describe '#run' do
     it 'splits a multilayer KML into single-layer KML' do
       source_file = SourceFile.new(@multiple_layer_filepath)
-      splitter    = KmlSplitter.new([source_file], @temporary_directory)
+      splitter    = KmlSplitter.new(source_file, @temporary_directory)
       splitter.run
-      splitter.source_files.length.must_equal 10
+      splitter.source_files.length.should eq 10
     end
   end
 
   describe '#layers_in' do
     it 'returns all layers name in the file' do
       source_file = SourceFile.new(@one_layer_filepath)
-      splitter    = KmlSplitter.new([source_file], @temporary_directory)
-      splitter.layers_in(source_file).must_equal ["Absolute_and_Relative"]
+      splitter    = KmlSplitter.new(source_file, @temporary_directory)
+      splitter.layers_in(source_file).should eq ["Absolute_and_Relative"]
 
       source_file = SourceFile.new(@multiple_layer_filepath)
-      splitter    = KmlSplitter.new([source_file], @temporary_directory)
-      splitter.layers_in(source_file).length.must_equal 10 
+      splitter    = KmlSplitter.new(source_file, @temporary_directory)
+      splitter.layers_in(source_file).length.should eq 10
     end
   end
 
@@ -37,6 +37,6 @@ describe KmlSplitter do
     File.expand_path(
       File.join(File.dirname(__FILE__), "../fixtures/#{filepath}")
     )
-  end #path_to
+  end
 end